+++ lib/k8s/application.ex
@@ -6,7 +6,8 @@ defmodule K8s.Application do
   def start(_type, _args) do
     children = [
       K8s.Client.Mint.ConnectionRegistry,
-      {DynamicSupervisor, name: K8s.Client.Mint.ConnectionSupervisor, strategy: :one_for_one}
+      {DynamicSupervisor, name: K8s.Client.Mint.ConnectionSupervisor, strategy: :one_for_one},
+      {DynamicSupervisor, name: K8s.Conn.Auth.ProviderSupervisor, strategy: :one_for_one}
     ]
 
     opts = [strategy: :one_for_one, name: K8s.Supervisor]